+++ crates/vm/levm/src/call_frame.rs
@@ -125,6 +125,31 @@ impl Stack {
         Ok(())
     }
 
+    #[inline]
+    pub fn push_zero(&mut self) -> Result<(), ExceptionalHalt> {
+        // Since the stack grows downwards, when an offset underflow is detected the stack is
+        // overflowing.
+        let next_offset = self
+            .offset
+            .checked_sub(1)
+            .ok_or(ExceptionalHalt::StackOverflow)?;
+
+        // The following index cannot fail because `next_offset` has already been checked and
+        // `self.offset` is known to be within `STACK_LIMIT`.
+        #[expect(unsafe_code, reason = "next_offset == self.offset - 1 >= 0")]
+        unsafe {
+            *self
+                .values
+                .get_unchecked_mut(next_offset)
+                .0
+                .as_mut_ptr()
+                .cast() = [0u64; 4];
+        }
+        self.offset = next_offset;
+
+        Ok(())
+    }
+
     pub fn len(&self) -> usize {
         // The following operation cannot underflow because `self.offset` is known to be less than
         // or equal to `self.values.len()` (aka. `STACK_LIMIT`).
@@ -303,6 +328,7 @@ impl CallFrame {
         memory: Memory,
     ) -> Self {
         // Note: Do not use ..Default::default() because it has runtime cost.
+
         Self {
             gas_limit,
             gas_remaining: gas_limit,
@@ -330,13 +356,14 @@ impl CallFrame {
 
     #[inline(always)]
     pub fn next_opcode(&self) -> u8 {
-        // 0 is the opcode stop.
-        self.bytecode.get(self.pc).copied().unwrap_or(0)
-    }
-
-    pub fn increment_pc_by(&mut self, count: usize) -> Result<(), VMError> {
-        self.pc = self.pc.checked_add(count).ok_or(InternalError::Overflow)?;
-        Ok(())
+        if self.pc < self.bytecode.len() {
+            #[expect(unsafe_code, reason = "bounds checked above")]
+            unsafe {
+                *self.bytecode.get_unchecked(self.pc)
+            }
+        } else {
+            0
+        }
     }
 
     pub fn pc(&self) -> usize {
@@ -429,7 +456,13 @@ impl<'a> VM<'a> {
         Ok(())
     }
 
-    pub fn increment_pc_by(&mut self, count: usize) -> Result<(), VMError> {
-        self.current_call_frame.increment_pc_by(count)
+    #[inline(always)]
+    pub fn advance_pc(&mut self, count: usize) -> Result<(), VMError> {
+        self.current_call_frame.pc = self
+            .current_call_frame
+            .pc
+            .checked_add(count)
+            .ok_or(InternalError::Overflow)?;
+        Ok(())
     }
 }