+++ crates/vm/levm/src/opcode_handlers/block.rs
@@ -5,7 +5,7 @@ use crate::{
     utils::*,
     vm::VM,
 };
-use ethrex_common::{U256, types::Fork, utils::u256_from_big_endian_const};
+use ethrex_common::utils::u256_from_big_endian_const;
 
 // Block Information (11)
 // Opcodes: BLOCKHASH, COINBASE, TIMESTAMP, NUMBER, PREVRANDAO, GASLIMIT, CHAINID, SELFBALANCE, BASEFEE, BLOBHASH, BLOBBASEFEE
@@ -23,7 +23,7 @@ impl<'a> VM<'a> {
         if block_number < current_block.saturating_sub(LAST_AVAILABLE_BLOCK_LIMIT)
             || block_number >= current_block
         {
-            current_call_frame.stack.push1(U256::zero())?;
+            current_call_frame.stack.push_zero()?;
             return Ok(OpcodeResult::Continue { pc_increment: 1 });
         }
 
@@ -138,10 +138,6 @@ impl<'a> VM<'a> {
     // BLOBHASH operation
     /// Currently not tested
     pub fn op_blobhash(&mut self) -> Result<OpcodeResult, VMError> {
-        // [EIP-4844] - BLOBHASH is only available from CANCUN
-        if self.env.config.fork < Fork::Cancun {
-            return Err(ExceptionalHalt::InvalidOpcode.into());
-        }
         self.current_call_frame
             .increase_consumed_gas(gas_cost::BLOBHASH)?;
         let index = self.current_call_frame.stack.pop1()?;
@@ -150,7 +146,7 @@ impl<'a> VM<'a> {
         let index = match u256_to_usize(index) {
             Ok(index) if index < blob_hashes.len() => index,
             _ => {
-                self.current_call_frame.stack.push1(U256::zero())?;
+                self.current_call_frame.stack.push_zero()?;
                 return Ok(OpcodeResult::Continue { pc_increment: 1 });
             }
         };
@@ -167,11 +163,6 @@ impl<'a> VM<'a> {
 
     // BLOBBASEFEE operation
     pub fn op_blobbasefee(&mut self) -> Result<OpcodeResult, VMError> {
-        // [EIP-7516] - BLOBBASEFEE is only available from CANCUN
-        if self.env.config.fork < Fork::Cancun {
-            return Err(ExceptionalHalt::InvalidOpcode.into());
-        }
-
         self.current_call_frame
             .increase_consumed_gas(gas_cost::BLOBBASEFEE)?;